final touches

+++ src/controller/CharaktereController.java
@@ -20,6 +20,12 @@ public class CharaktereController implements IController<Charaktere, Integer> {
                 .toList();
     }
 
+    /**
+     * Diese filtriert nach ein Benutzer input die produkte die diesen UNiversum haben
+     * @param x
+     * @return die liste der Charakteren
+     */
+
     public List<Charaktere> aufgabe4(String x) {
         return getAll().stream()
                 .filter(charaktere -> charaktere.getProdukte().stream()
@@ -32,9 +38,9 @@ public class CharaktereController implements IController<Charaktere, Integer> {
 
         List<Produkte> produkte = selectedCharaktere.getProdukte();
         if(sort == 1) {
-            produkte.sort(Comparator.comparingInt(Produkte::getPreis));
+            produkte.sort(Comparator.comparingDouble(Produkte::getPreis));
         } else {
-            produkte.sort((p1, p2) -> Integer.compare(p2.getPreis(), p1.getPreis()));
+            produkte.sort((p1, p2) -> Double.compare(p2.getPreis(), p1.getPreis()));
         }
         return produkte;
     }
